Fix: Update Dockerfile to use WAR file
The Dockerfile was updated to use a Tomcat base image and copy the WAR file instead of the JAR file. The project was updated to generate a WAR file, so the Dockerfile needed to be updated to reflect this change.

The README.md file was also updated to include a section on how to build and run the application using Docker.