Add wrestler decks and cards (#13)
* Fix: Update Dockerfile to use WAR file

The Dockerfile was updated to use a Tomcat base image and copy the WAR file instead of the JAR file. The project was updated to generate a WAR file, so the Dockerfile needed to be updated to reflect this change.

The README.md file was also updated to include a section on how to build and run the application using Docker.

* Fix: Add dependency-check-maven plugin to pom.xml

The dependency-check-maven plugin was added to the pom.xml file to align with the README.md file.

* Fix(maven.yml): Update Docker image build commands to include production profile and skip tests

* Add new cards and wrestler set for Brian Pillman

* Add new cards and wrestler set for Matt Cardona

* Add new cards and wrestler set for Gilgamesh

* Add new cards and wrestler set for Kamala

* Update card amount for ATW set and remove duplicate entry

* Add new wrestler and card set for Johnny All Time

* Add new cards and wrestler set for André the Giant

* Add new wrestler Mukundi Shumba and card set with moves

* Add new wrestler Kurt Angle and card set with moves